fix server crash
